pyproject.toml
[tool.poetry]
name = "dio-bank"
version = "0.1.0"
description = ""
authors = ["allanapavao <allanarodrigues.p@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.1.0"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"
black = "^24.10.0"
flask-jwt-extended = "^4.7.1"
psycopg2-binary = "^2.9.10"
gunicorn = "^23.0.0"
flask-bcrypt = "^1.0.1"
flask-marshmallow = "^1.3.0"
marshmallow-sqlalchemy = "^1.4.0"
 
[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-mock = "^3.14.0"

[tool.pytest.ini_options]
addopts = "-vvv"
testpaths = [
    "tests",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
